TEST-31 Relocated maven utiloities for reuse in downstream projects
