+++ dominion-pax-exam-driver/src/main/java/org/codice/dominion/pax/exam/options/karaf/extensions/FeatureExtension.java
@@ -17,7 +17,7 @@ import org.codice.dominion.options.Options.MavenUrl;
 import org.codice.dominion.options.karaf.KarafOptions.Feature;
 import org.codice.dominion.pax.exam.interpolate.PaxExamInterpolator;
 import org.codice.dominion.pax.exam.options.PaxExamOption.Extension;
-import org.codice.dominion.pax.exam.options.extensions.Utilities;
+import org.codice.dominion.pax.exam.options.Utilities;
 import org.codice.dominion.resources.ResourceLoader;
 import org.ops4j.pax.exam.Option;
 import org.ops4j.pax.exam.karaf.options.KarafDistributionOption;
